# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CrossBorder AI Demo API",
    "version": "1.3.0"
  },
  "paths": {
...
    "/api/demo/context": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取香港当前时间信息",
        "description": "返回香港当前时间、有效预测范围和轮询间隔。",
        "operationId": "get_demo_context_api_demo_context_get",
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoContextResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/personas": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取本地 Demo 身份",
        "operationId": "get_demo_personas_api_demo_personas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoPersonasResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/v1-model": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取 AI v1 合成数据评估与产物状态",
        "operationId": "get_v1_model_api_demo_v1_model_get",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ModelResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/v2-model": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取 AI v2.2 透明校准课堂模型状态",
        "operationId": "get_v2_model_api_demo_v2_model_get",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ModelResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/v1-readiness": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取 V1 完整 Demo 就绪度",
        "operationId": "get_v1_readiness_api_demo_v1_readiness_get",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ReadinessResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/audit": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取本地 Demo 写操作审计",
        "operationId": "get_audit_events_api_demo_audit_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventListResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/operations-summary": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取本地 Demo 运营分析汇总",
        "operationId": "get_operations_summary_api_demo_operations_summary_get",
        "parameters": [
          {
            "name": "window_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 24,
              "title": "Window Hours"
            }
          },
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsSummaryResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/model-shadow-summary": {
      "get": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "获取 AI v1 影子观测汇总",
        "operationId": "get_model_shadow_summary_api_demo_model_shadow_summary_get",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShadowObservationSummaryResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/demo/reset": {
      "post": {
        "tags": [
          "Demo 控制"
        ],
        "summary": "重置 Demo 动态数据",
        "description": "删除已保存的反馈、订阅和企业方案，然后重新加载 JSON 种子。",
        "operationId": "reset_demo_api_demo_reset_post",
        "parameters": [
          {
            "name": "X-Demo-Persona-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。",
              "title": "X-Demo-Persona-Id"
            },
            "description": "本地课堂 Demo 身份；受保护接口缺失时返回 401，不是生产认证令牌。"
          }
        ],
        "responses": {
          "200": {
            "description": "重置成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoResetResponse"
                }
              }
            }
          },
          "404": {
            "description": "请求的资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "请求与当前状态冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "领域规则或请求参数验证失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "内部服务或持久化错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEventListResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "events",
          "total"
        ],
        "title": "AuditEventListResponse"
      },
...
      "DemoContextResponse": {
        "properties": {
          "current_time": {
            "type": "string",
            "format": "date-time",
            "title": "Current Time"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "min_target_time": {
            "type": "string",
            "format": "date-time",
            "title": "Min Target Time"
          },
          "suggested_target_time": {
            "type": "string",
            "format": "date-time",
            "title": "Suggested Target Time"
          },
          "max_target_time": {
            "type": "string",
            "format": "date-time",
            "title": "Max Target Time"
          },
          "poll_interval_seconds": {
            "type": "integer",
            "title": "Poll Interval Seconds"
          }
        },
        "type": "object",
        "required": [
          "current_time",
          "timezone",
          "min_target_time",
          "suggested_target_time",
          "max_target_time",
          "poll_interval_seconds"
        ],
        "title": "DemoContextResponse"
      },
...
      "DemoPersonasResponse": {
        "properties": {
          "default_persona_id": {
            "type": "string",
            "title": "Default Persona Id"
          },
          "personas": {
            "items": {
              "$ref": "#/components/schemas/DemoPersona"
            },
            "type": "array",
            "title": "Personas"
          }
        },
        "type": "object",
        "required": [
          "default_persona_id",
          "personas"
        ],
        "title": "DemoPersonasResponse"
      },
      "DemoResetResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "seeded": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Seeded"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "seeded",
          "message"
        ],
        "title": "DemoResetResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "OperationsSummaryResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "window_hours": {
            "type": "integer",
            "title": "Window Hours"
          },
          "forecast": {
            "additionalProperties": true,
            "type": "object",
            "title": "Forecast"
          },
          "crowdsource": {
            "additionalProperties": true,
            "type": "object",
            "title": "Crowdsource"
          },
          "errors": {
            "additionalProperties": true,
            "type": "object",
            "title": "Errors"
          },
          "audit": {
            "additionalProperties": true,
            "type": "object",
            "title": "Audit"
          },
          "adapters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Adapters"
          },
          "commercial": {
            "additionalProperties": true,
            "type": "object",
            "title": "Commercial"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "window_hours",
          "forecast",
          "crowdsource",
          "errors",
          "audit",
          "adapters",
          "commercial"
        ],
        "title": "OperationsSummaryResponse"
      },
...
      "ShadowObservationSummaryResponse": {
        "properties": {
          "total_observations": {
            "type": "integer",
            "title": "Total Observations"
          },
          "available_observations": {
            "type": "integer",
            "title": "Available Observations"
          },
          "unavailable_observations": {
            "type": "integer",
            "title": "Unavailable Observations"
          },
          "latest_observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Observed At"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/ShadowObservationPortSummary"
            },
            "type": "array",
            "title": "Ports"
          }
        },
        "type": "object",
        "required": [
          "total_observations",
          "available_observations",
          "unavailable_observations",
          "ports"
        ],
        "title": "ShadowObservationSummaryResponse"
      },
...
      "V1ModelResponse": {
        "properties": {
          "artifact_available": {
            "type": "boolean",
            "title": "Artifact Available"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "synthetic_only": {
            "type": "boolean",
            "title": "Synthetic Only"
          },
          "dataset": {
            "additionalProperties": true,
            "type": "object",
            "title": "Dataset"
          },
          "split": {
            "additionalProperties": true,
            "type": "object",
            "title": "Split"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "promotion": {
            "additionalProperties": true,
            "type": "object",
            "title": "Promotion"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "artifact_available",
          "model_version",
          "synthetic_only",
          "dataset",
          "split",
          "metrics",
          "features",
          "promotion",
          "limitations"
        ],
        "title": "V1ModelResponse"
      },
      "V1ReadinessCheck": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "name",
          "passed",
          "detail"
        ],
        "title": "V1ReadinessCheck"
      },
      "V1ReadinessResponse": {
        "properties": {
          "demo_ready": {
            "type": "boolean",
            "title": "Demo Ready"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/V1ReadinessCheck"
            },
            "type": "array",
            "title": "Checks"
          },
          "adapter_modes": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Adapter Modes"
          }
        },
        "type": "object",
        "required": [
          "demo_ready",
          "checks",
          "adapter_modes"
        ],
        "title": "V1ReadinessResponse"
      },
      "V2ModelResponse": {
        "properties": {
          "artifact_available": {
            "type": "boolean",
            "title": "Artifact Available"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "synthetic_only": {
            "type": "boolean",
            "title": "Synthetic Only"
          },
          "evaluation_scope": {
            "type": "string",
            "title": "Evaluation Scope"
          },
          "dataset": {
            "additionalProperties": true,
            "type": "object",
            "title": "Dataset"
          },
          "split": {
            "additionalProperties": true,
            "type": "object",
            "title": "Split"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          },
          "target_scope": {
            "type": "string",
            "title": "Target Scope"
          },
          "real_feature_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Real Feature Sources"
          },
          "calibration_version": {
            "type": "string",
            "title": "Calibration Version"
          },
          "source_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Snapshot"
          },
          "data_audit": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data Audit"
          },
          "formula": {
            "additionalProperties": true,
            "type": "object",
            "title": "Formula"
          },
          "selection": {
            "additionalProperties": true,
            "type": "object",
            "title": "Selection"
          },
          "candidate_leaderboard": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Candidate Leaderboard"
          },
          "interval_calibration": {
            "additionalProperties": true,
            "type": "object",
            "title": "Interval Calibration"
          },
          "traffic_distribution": {
            "additionalProperties": true,
            "type": "object",
            "title": "Traffic Distribution"
          },
          "sensitivity": {
            "additionalProperties": true,
            "type": "object",
            "title": "Sensitivity"
          },
          "promotion": {
            "additionalProperties": true,
            "type": "object",
            "title": "Promotion"
          },
          "optimization_matrix": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Optimization Matrix"
          }
        },
        "type": "object",
        "required": [
          "artifact_available",
          "model_version",
          "synthetic_only",
          "evaluation_scope",
          "dataset",
          "split",
          "metrics",
          "features",
          "limitations",
          "target_scope",
          "real_feature_sources",
          "calibration_version",
          "source_snapshot",
          "data_audit",
          "formula",
          "selection",
          "candidate_leaderboard",
          "interval_calibration",
          "traffic_distribution",
          "sensitivity",
          "promotion",
          "optimization_matrix"
        ],
        "title": "V2ModelResponse"
      },
...
    }
  }
}